+++ rules/nette-kdyby/tests/Rector/ClassMethod/ReplaceMagicPropertyWithEventClassRector/Source/Tomato.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\NetteKdyby\Tests\Rector\ClassMethod\ReplaceMagicPropertyWithEventClassRector\Source;
+
+final class Tomato
+{
+    public function unwrap()
+    {
+    }
+}
